Added constraint to see if the database is available before testing for online updates, messages and getting/setting system preferences.
This stops spurious error messages from appearing in the log when the initial database creation is performed.

+++ branches/development/swisscenter/base/prefs.php
@@ -22,7 +22,8 @@
  
  function get_sys_pref( $pref, $default = '' )
  {
-   $result = db_value("select value from system_prefs where name='".strtoupper($pref)."'");
+   if ( test_db() == 'OK' );
+     $result = db_value("select value from system_prefs where name='".strtoupper($pref)."'");
 
    if ($result == '')
      return $default;
@@ -52,7 +53,7 @@
  function set_sys_pref( $name, $value)
  {
    // Only update if the value changes
-   if (db_value("select count(*) from system_prefs where name='".strtoupper($name)."' and value='$value'") == 0)
+   if (test_db() == 'OK' && db_value("select count(*) from system_prefs where name='".strtoupper($name)."' and value='$value'") == 0)
    {
      db_sqlcommand("delete from system_prefs where name='".strtoupper($name)."'");
      $result = db_insert_row('system_prefs', array("NAME"=>strtoupper($name), "VALUE"=>$value) );
